+++ lib/flexible_tree_view.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2022, the tree_view project authors. Please see the AUTHORS file
+// Copyright (c) 2022, the flexible_tree_view project authors. Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.