+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.0.2
+
+- Add `TreeNode.hasNodes` method
+- Add `TreeNode.addNodes` method
+- Add `TreeNode.clearNodes` method
+- Add `TreeNode.nodeIndexAt` method
+- Add `TreeNode.removeNodes` method
+- Add `TreeNode.removeNodeAt` method
+- Add `TreeNode.removeRange` method
+- Add `TreeNode.removeWhere` method
+
 ## 0.0.1
 
 Initial project.